@@ -1,16 +1,18 @@
 buildscript {
     repositories {
+        google()
         jcenter()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.2'
+        classpath 'com.android.tools.build:gradle:3.2.0'
     }
 }
 
 allprojects {
     repositories {
         jcenter()
+        google()
         maven {
             url "https://maven.google.com"
         }
@@ -19,8 +21,8 @@ allprojects {
 
 ext {
     minSdkVersion = 14
-    compileSdkVersion = 24
-    buildToolsVersion = '24.0.3'
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.3'
 
     junitVersion = '4.12'
     mockitoVersion = '1.10.19'
