@@ -15,8 +15,8 @@ android {
         disable 'InvalidPackage'
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     dexOptions {
@@ -25,16 +25,17 @@ android {
 }
 
 dependencies {
-    compile project(':scissors2')
-    compile 'com.android.support:appcompat-v7:' + rootProject.ext.supportVersion
-    compile 'com.android.support:design:' + rootProject.ext.supportVersion
-    compile 'com.jakewharton:butterknife:7.0.1'
-    compile 'com.squareup.leakcanary:leakcanary-android:1.3.1'
-    compile 'com.squareup.picasso:picasso:2.5.2'
+    implementation project(':scissors2')
+    implementation 'com.android.support:appcompat-v7:' + rootProject.ext.supportVersion
+    implementation 'com.android.support:design:' + rootProject.ext.supportVersion
+    implementation 'com.jakewharton:butterknife:10.0.0'
+    implementation 'com.squareup.leakcanary:leakcanary-android:1.3.1'
+    implementation 'com.squareup.picasso:picasso:2.5.2'
+    implementation 'com.google.android.material:material:1.1.0-alpha06'
     // Or Glide
     // compile 'com.github.bumptech.glide:glide:3.6.1'
     // Or Android-Universal-Image-Loader
     // compile 'com.nostra13.universalimageloader:universal-image-loader:1.9.4'
-    compile 'io.reactivex:rxjava:1.0.15'
-    compile 'io.reactivex:rxandroid:1.0.1'
+    implementation 'io.reactivex:rxjava:1.0.15'
+    implementation 'io.reactivex:rxandroid:1.0.1'
 }
