@@ -15,8 +15,8 @@
  */
 package com.lyft.android.scissors2;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.widget.ImageView;
 
 /**
