@@ -1,7 +1,7 @@
 package com.lyft.android.scissors2;
 
 import android.graphics.Bitmap;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.File;
 import java.io.OutputStream;
