@@ -1,7 +1,7 @@
 package com.lyft.android.scissors2;
 
 import android.graphics.Bitmap;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.ViewTreeObserver;
 
 import static com.lyft.android.scissors2.CropView.Extensions.LoaderType;
